fix(ng-add): undefined project in assets and tsconfig paths
