+++ libs/add/schematics/ng-add/testing/ng-add-standalone-app.spec.ts
@@ -88,10 +88,10 @@ describe('ng-add standalone app', () => {
   "compilerOptions": {
     "paths": {
       "@ng-doc/generated": [
-        "./ng-doc//index.ts"
+        "./ng-doc/demo/index.ts"
       ],
       "@ng-doc/generated/*": [
-        "./ng-doc//*"
+        "./ng-doc/demo/*"
       ]
     }
   }
@@ -196,7 +196,7 @@ export class AppComponent {
                 },
                 {
                   "glob": "**/*",
-                  "input": "ng-doc//assets",
+                  "input": "ng-doc/demo/assets",
                   "output": "assets/ng-doc"
                 }
               ],
