+++ libs/add/schematics/ng-add/utils/get-project.ts
@@ -1,6 +1,7 @@
 import { ProjectDefinition, WorkspaceDefinition } from '@angular-devkit/core/src/workspace';
 
 import { Schema } from '../schema';
+import { getProjectName } from './get-project-name';
 
 /**
  *
@@ -11,14 +12,7 @@ export function getProject(
 	options: Schema,
 	workspace: WorkspaceDefinition,
 ): ProjectDefinition | undefined {
-	const firstBuildableProjectName: string | undefined = Array.from(
-		workspace.projects.entries(),
-	).find(([, project]: [string, ProjectDefinition]) => project.targets.get('build'))?.[0];
+	const projectName = getProjectName(options, workspace);
 
-	const projectName: string =
-		options.project ||
-		workspace.extensions['defaultProject']?.toString() ||
-		firstBuildableProjectName ||
-		'';
 	return workspace.projects.get(projectName);
 }
